requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
sentence-transformers==5.1.0
numpy==2.1.3
pandas==2.3.2
faiss-cpu==1.12.0
networkx==3.4.2
SQLAlchemy==2.0.43
python-dotenv==1.1.1
pymongo==4.14.1
redis==6.4.0
neo4j==5.28.2
